requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Bcrypt==0.5.2
Flask-Login==0.2.7
Flask-Script==0.6.3
Flask-SQLAlchemy==1.0
beautifulsoup4==4.3.2
boto==2.17.0
closure==20121212
gevent==0.13.8
ghdiff==0.1
gittle==0.2.2
itsdangerous==0.23
lxml==3.2.4
markdown2==2.1.0
pyzmq==14.0.0
recaptcha==1.0rc1
recaptcha-client==1.0.6
redis==2.8.0
simplejson==3.3.1
sockjs-tornado==1.0.0
supervisor==3.0
SQLAlchemy==0.8.3
tornado==3.1.1
tldextract==1.2.2
psycopg2==2.5.1
